requirements.txt
setuptools~=68.0.0
pynput~=1.7.6
psutil~=5.9.5
h5py>=3.10.0
numpy~=1.22
matplotlib~=3.7.2
Pillow~=10.0.0
mss~=9.0.1
scikit-learn~=1.3.0
git+https://github.com/DLR-RM/stable-baselines3@master#egg=stable-baselines3
pyautogui~=0.9.54
scikit-image~=0.21.0
gymnasium~=0.29.0
tensorboard~=2.16.2
tensorboard-plugin-profile~=2.15.1
tensorflow~=2.16.1
git+https://github.com/Stable-Baselines-Team/stable-baselines3-contrib@master#egg=sb3-contrib
pytesseract~=0.3.10
opencv-python